+++ changelog.md
@@ -2,6 +2,9 @@ Codion Change Log
 ==================
 
 ## 0.17.21-SNAPSHOT
+### is.codion.swing.framework.ui
+- EntityPanel bug fixed, table refresh control was always enabled, instead of only when the panel was active.
+- EntityPanel.createControls() added, related refactoring.
 
 ## 0.17.20
 ### is.codion.common.model